refactor: :recycle: 라우팅 / 레이아웃 수정
