+++ src/components/Layout/Header.jsx
@@ -2,12 +2,13 @@ import { Divider, HStack, Spacer, Text } from '@chakra-ui/react';
 import { useNavigate } from 'react-router-dom';
 import ControllBar from './ControllBar';
 
-const Header = () => {
+const Header = ({ children }) => {
   const navigate = useNavigate();
   return (
     <>
       <HStack>
-        <Text onClick={() => navigate('/')}>헤더입니다</Text>
+        <Text onClick={() => navigate('/')}>메인 헤더</Text>
+        {children}
         <Spacer />
         <ControllBar />
       </HStack>
