+++ src/components/Layout/MainLayout.jsx
@@ -1,0 +1,14 @@
+import { Outlet } from 'react-router-dom';
+import Footer from './Footer';
+import Header from './Header';
+
+const MainLayout = () => {
+  return (
+    <>
+      <Header />
+      <Outlet />
+      <Footer />
+    </>
+  );
+};
+export default MainLayout;
